[REF] {hr_,}gamification: simplify code structure
Before this commit, the code structure was set uselessly complicated
in `hr_gamification` while not bringing any additional value.
This commit:
- greatly simplifies the hr_gamification code
- moves the count smartbutton on badges from hr_gamification
to gamification while making the count related to users rather than
employees
- removes the link between gamification_badge_user; a badge cannot
be set only for one of the user's employee and will always be shared.

task-4247122
