+++ addons/hr_gamification/models/__init__.py
@@ -1,6 +1,4 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import gamification
 from . import hr_employee
 from . import res_users
