+++ addons/hr_gamification/models/hr_employee.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import api, fields, models
@@ -8,15 +7,11 @@ class HrEmployeeBase(models.AbstractModel):
     _inherit = "hr.employee.base"
 
     goal_ids = fields.One2many('gamification.goal', string='Employee HR Goals', compute='_compute_employee_goals')
+    has_badges = fields.Boolean(compute='_compute_employee_badges')
     badge_ids = fields.One2many(
         'gamification.badge.user', string='Employee Badges', compute='_compute_employee_badges',
         help="All employee badges, linked to the employee either directly or through the user"
     )
-    has_badges = fields.Boolean(compute='_compute_employee_badges')
-    # necessary for correct dependencies of badge_ids and has_badges
-    direct_badge_ids = fields.One2many(
-        'gamification.badge.user', 'employee_id',
-        help="Badges directly linked to the employee")
 
     @api.depends('user_id.goal_ids.challenge_id.challenge_category')
     def _compute_employee_goals(self):
@@ -26,13 +21,14 @@ class HrEmployeeBase(models.AbstractModel):
                 ('challenge_id.challenge_category', '=', 'hr'),
             ])
 
-    @api.depends('direct_badge_ids', 'user_id.badge_ids.employee_id')
+    @api.depends('user_id.badge_ids')
     def _compute_employee_badges(self):
+        badge_read_group = self.env['gamification.badge.user']._read_group(
+            domain=[('user_id', 'in', self.user_id.ids)],
+            groupby=['user_id'],
+            aggregates=['id:recordset'])
+        badges_per_user = dict(badge_read_group)
         for employee in self:
-            badge_ids = self.env['gamification.badge.user'].search([
-                '|', ('employee_id', 'in', employee.ids),
-                     '&', ('employee_id', '=', False),
-                          ('user_id', 'in', employee.user_id.ids)
-            ])
-            employee.has_badges = bool(badge_ids)
-            employee.badge_ids = badge_ids
+            employee_badges = badges_per_user.get(employee.user_id, self.env['gamification.badge.user'])
+            employee.has_badges = bool(employee_badges)
+            employee.badge_ids = employee_badges
