+++ addons/hr_gamification/views/menuitems.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+        <menuitem id="menu_hr_gamification"
+            parent="hr.menu_human_resources_configuration"
+            name="Challenges" sequence="100"/>
+
+            <menuitem id="gamification_badge_menu_hr"
+                parent="menu_hr_gamification"
+                action="gamification.badge_list_action" />
+
+            <menuitem id="gamification_challenge_menu_hr"
+                parent="menu_hr_gamification"
+                action="challenge_list_action2"
+                groups="hr.group_hr_user"/>
+
+            <menuitem id="gamification_goal_menu_hr"
+                parent="menu_hr_gamification"
+                action="goals_menu_groupby_action2"
+                groups="hr.group_hr_user"/>
+</odoo>
